Split widgets: performance list and row
