+++ extension/devtools/app/lib/main.dart
@@ -55,6 +55,8 @@ part 'widgets/collections_panel.dart';
 part 'widgets/batching_panel.dart';
 part 'widgets/timeline_panel.dart';
 part 'widgets/performance_panel.dart';
+part 'widgets/performance_list.dart';
+part 'widgets/performance_row.dart';
 part 'widgets/computed_header.dart';
 part 'widgets/computed_list.dart';
 part 'widgets/computed_table_header.dart';
@@ -222,94 +224,6 @@ class OrefPalette {
   static const Color pink = Color(0xFFFF71C6);
   static const Color deepBlue = Color(0xFF0C141C);
   static const Color lightBlue = Color(0xFFE7F3FF);
-}
-
-class _PerformanceList extends StatelessWidget {
-  const _PerformanceList({required this.samples});
-
-  final List<UiPerformanceSample> samples;
-
-  @override
-  Widget build(BuildContext context) {
-    return _GlassCard(
-      padding: const EdgeInsets.all(0),
-      child: samples.isEmpty
-          ? Padding(
-              padding: const EdgeInsets.all(16),
-              child: Text(
-                'No performance samples yet.',
-                style: Theme.of(context).textTheme.bodyMedium,
-              ),
-            )
-          : Padding(
-              padding: const EdgeInsets.all(16),
-              child: Column(
-                children: [
-                  for (var index = 0; index < samples.length; index++) ...[
-                    _PerformanceRow(sample: samples[index]),
-                    if (index != samples.length - 1) const SizedBox(height: 12),
-                  ],
-                ],
-              ),
-            ),
-    );
-  }
-}
-
-class _PerformanceRow extends StatelessWidget {
-  const _PerformanceRow({required this.sample});
-
-  final UiPerformanceSample sample;
-
-  @override
-  Widget build(BuildContext context) {
-    final textTheme = Theme.of(context).textTheme;
-    return _GlassCard(
-      padding: const EdgeInsets.all(16),
-      child: DefaultTextStyle.merge(
-        style: textTheme.bodyMedium,
-        child: Row(
-          children: [
-            Expanded(
-              flex: 2,
-              child: Text(
-                _formatAge(sample.timestamp),
-                textAlign: TextAlign.center,
-              ),
-            ),
-            Expanded(
-              flex: 2,
-              child: Text(
-                '${sample.signalWrites} writes',
-                textAlign: TextAlign.center,
-              ),
-            ),
-            Expanded(
-              flex: 2,
-              child: Text(
-                '${sample.effectRuns} runs',
-                textAlign: TextAlign.center,
-              ),
-            ),
-            Expanded(
-              flex: 2,
-              child: Text(
-                _formatDurationUs(sample.avgEffectDurationUs.round()),
-                textAlign: TextAlign.center,
-              ),
-            ),
-            Expanded(
-              flex: 2,
-              child: Text(
-                '${sample.collectionMutations} mutations',
-                textAlign: TextAlign.center,
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
 }
 
 class _SettingsPanel extends StatelessWidget {
